Minimal port of better-apidoc to an extension
This is a basic proof-of-concept port of
https://github.com/goerz/better-apidoc, exploring how intrusive
templating is to the existing code of sphinxcontrib.apidoc. The code has
been tested (manually) on the master branch of
https://github.com/mabuchilab/QNET
to run through and generate the same results at the original
better-apidoc.

The only configuration that was required in QNET's docs/conf.py:

    # apidoc settings
    apidoc_module_dir = '../src/qnet'
    apidoc_output_dir = './API'
    apidoc_separate_modules = True
    apidoc_templates = templates_path[0]

There is no documentation or testing for the added templating
functionality in sphinxcontrib.apidoc at this time.